week-06/seminar/consul-demo/app/requirements.txt
python-consul==1.1.0
Flask==2.2.2
Werkzeug==2.2.2
